+++ pcc_colonel_blotto/__init__.py
@@ -1,3 +1,3 @@
 """PCC Colonel Blotto."""
 
-__version__ = "0.1.0"
+__version__ = "0.4.0"